requirements.txt
groq
python-dotenv
requests
pyyaml
plotext